requirements.txt
tqdm
numpy
pandas
librosa
torch
openai-whisper
transformers
jiwer
demucs
soundfile
numba
pypinyin
opencc-python-reimplement
chinese_converter